refactor(STADTPULS-645): Restructure URL to include "sensors" in route
