+++ src/components/SensorPageHeader/index.tsx
@@ -54,7 +54,7 @@ const UserLink: FC<Pick<ParsedSensorType, "authorName" | "authorUsername">> = ({
   authorName,
   authorUsername,
 }) => (
-  <Link href={`/${authorUsername}`}>
+  <Link href={`/${authorUsername}/sensors`}>
     <a
       className={[
         "flex gap-2 items-center leading-tight",
